Use `llvm.dbg.value` instead of `llvm.dbg.declare` for SSA values
Also insert the debug instrinsic just after the instruction that
defines it.

Technically these are not addresses in memory, so it's important to
use `dbg.value` instead of `dbg.declare`.

Without this change, almost none of the local SSA values are
available in gdb (it might be just the arguments that are) but
with it I can very often print the arguments of the instruction
execution is halted on.